// src/types/pokemon.ts
export interface Species {
    name: string;
    url:  string;
}
// ...
export interface Type {
    slot: number;
    type: Species;
}